New way to pass in the animator for each transition. 🚀
This should allow for any caller to pass a custom animator instead of centralizing everything through the delegate object.
