+++ Sources/Extensions.swift
@@ -23,10 +23,6 @@ extension UIViewAutoresizing {
 	}
 }
 
-prefix func ! <T>(closure: @escaping (T) -> Bool) -> (T) -> Bool {
-	return { !closure($0) }
-}
-
 extension Bool {
 
 	///
